added more precise error handling to resolvers.js and fixed npm start issue
